services/ingest_worker/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
google-cloud-firestore==2.16.0
google-cloud-aiplatform==1.59.0
google-cloud-pubsub==2.19.7
